requirements.txt
bidict==0.23.1
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
eventlet==0.40.3
filelock==3.19.1
Flask==3.1.2
flask-cors==6.0.1
Flask-Login==0.6.3
Flask-SocketIO==5.5.1
fsspec==2025.7.0
greenlet==3.2.4
h11==0.16.0
huggingface-hub==0.34.4
idna==3.10
itsdangerous==2.2.0
jellyfish==1.2.0
Jinja2==3.1.6
joblib==1.5.2
MarkupSafe==3.0.2
mpmath==1.3.0
mysql-connector-python==9.4.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.6
packaging==25.0
python-dotenv==1.1.1
python-engineio==4.12.2
python-socketio==5.13.0
PyYAML==6.0.2
regex==2025.8.29
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.15.3
segtok==1.5.11
simple-websocket==1.1.0
sympy==1.14.0
tabulate==0.9.0
threadpoolctl==3.6.0
tokenizers==0.19.1
torch==2.8.0+cpu
tqdm==4.67.1
transformers==4.43.3
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
wsproto==1.2.0
yake==0.6.0
